requirements.txt
numpy
pandas
opencv-python
mediapipe
tensorflow
ultralytics
SpeechRecognition
pyttsx3
PyQt5
flet
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
jinja2
python-multipart
pydantic-settings
